Refactored getting href for sending email and fix send button on Team page
